# api/api-spec.yaml
openapi: 3.0.3
info:
  title: TripleThink API
  description: |
    Event-sourced narrative construction system for multi-book fiction series.

    TripleThink enables authors to manage complex storylines with:
    - Epistemic tracking (who knows what, when)
    - Fiction management (false narratives with strict audience constraints)
    - Time-travel queries (entity state at any timestamp)
    - Token-efficient AI integration

    ## Key Concepts

    - **Entities**: Characters, events, objects, locations, systems
    - **Metadata**: Rich author notes and AI guidance (loaded on-demand)
    - **Knowledge States**: What characters believe at specific times
    - **Fictions**: False narratives with target audiences
    - **Epistemic Queries**: The power feature for tracking beliefs
  version: 1.0.0
  contact:
    name: TripleThink Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/search:
    get:
      tags: [Search]
      summary: Full-text search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: date_range
          in: query
          schema:
            type: string
        - name: include_metadata
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Search results
  /api/search/relationships/{entityId}:
    get:
      tags: [Search]
      summary: Get related entities
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 1
            maximum: 3
        - name: relationship_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Related entities

  # ==========================================
  # AI (Token-Efficient Interface)
  # ==========================================
...
components:
  schemas:
    EntityType:
      type: string
      enum: [event, character, object, location, system, fiction]
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []